test: rewrite tests
